refactor(kernel): 优化数据类型定义和清理未使用变量

+++ main/kernel/process/process.c
@@ -60,7 +60,7 @@ void process_subsystem_init(void) {
 }
 
 static void free_pid(pid_t pid) {
-    if (pid == 0 || pid >= pid_capacity) {
+    if (pid == 0 || pid >= (int32_t)pid_capacity) {
         return;
     }
